Test für Stateklasse erstellt

+++ Parkhaussimulation/src/produktionscode/IF_State.java
@@ -2,6 +2,8 @@ package produktionscode;
 
 public interface IF_State {
 	
+	
+	
 	public Double getGesamtEinnahmen();
 	public Double getAvgEinnahmen();
 	public Double getAvgParkdauer();
@@ -14,7 +16,7 @@ public interface IF_State {
 	public void setGesamtEinnahmen(Double gesamtEinnahmen);
 	public void setAvgEinnahmen(Double avgEinnahmen);
 	public void setAvgParkdauer(Double avgParkdauer);
-	public void setBesucheranzahl(Double besucheranzahl);
+	public void setBesucheranzahl(int besucheranzahl);
 	public void setMinEinnahmen(Double minEinnahmen);
 	public void setMinParkdauer(Double minParkdauer);
 	public void setMaxEinnahmen(Double maxEinnahmen);
